decoupled webcontroller to home, search and useraccess and code refactoring
